remove unnecessary classes and some rename.
